Decouple compression, decompression and registry
